# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Geeks team app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of Geeks team app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://geeks-backend-project.onrender.com"
    }
  ],
...
  "paths": {
    "/articles": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Get list of articles",
        "operationId": "getArticles",
        "description": "Get list of articles with this endpoint",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "example": 12
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found articles!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found articles!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "articles",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "articles": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/articles"
                          }
                        },
                        "page": {
                          "type": "integer",
                          "example": 2
                        },
                        "perPage": {
                          "type": "integer",
                          "example": 4
                        },
                        "totalItems": {
                          "type": "integer",
                          "example": 20
                        },
                        "totalPages": {
                          "type": "integer",
                          "example": 5
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Articles"
        ],
        "summary": "Create an article",
        "operationId": "createArticle",
        "description": "Create an article with form-data including image upload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "article",
                  "img"
                ],
                "properties": {
                  "title": {
                    "description": "Article title",
                    "type": "string",
                    "example": "When anxiety fills the space: the path to peace"
                  },
                  "article": {
                    "description": "Contents of the article",
                    "type": "string",
                    "example": "In today`s world, where the rapid pace of events and the endless flow of information often rob us of our inner peace, finding harmony in everyday life becomes a real challenge"
                  },
                  "img": {
                    "description": "Upload article image (max 1MB)",
                    "type": "string",
                    "format": "binary"
                  },
                  "desc": {
                    "description": "Short description of the article",
                    "type": "string",
                    "example": "Meditations that help restore inner peace"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created an article",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created an article!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/articles"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get list of authors",
        "operationId": "getAuthors",
        "description": "Retrieve a paginated list of authors",
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer"
            },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found authors!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found authors!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/user"
                          }
                        },
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "perPage": {
                          "type": "number",
                          "example": 10
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 25
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 3
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/users/me/saved-articles/{articleId}": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Add article to saved list",
        "operationId": "addArticleToSaved",
        "description": "Add an article to the saved articles list of a specific user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "articleId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "66ff102bea7cc2e11f782401"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Article successfully added to saved list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully saved the article!"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Remove article from saved list",
        "operationId": "deleteArticleFromSaved",
        "description": "Remove an article from the saved articles list of a specific user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "articleId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "66ff102bea7cc2e11f782401"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully removed the article from saved list"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/users/me/subscriptions": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user subscriptions",
        "operationId": "getUserSubscriptions",
        "description": "Returns a list of users the authenticated user is subscribed to.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched subscriptions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully fetched subscriptions!"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/user"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/users/{authorId}/subscribe": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Subscribe to an author",
        "operationId": "subscribeToAuthor",
        "description": "Authenticated user subscribes to another user (author).",
        "parameters": [
          {
            "in": "path",
            "name": "authorId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the author to subscribe to"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully subscribed to author",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully subscribed to author!"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/user"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/users/{authorId}/unsubscribe": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Unsubscribe from an author",
        "operationId": "unsubscribeFromAuthor",
        "description": "Authenticated user unsubscribes from another user (author).",
        "parameters": [
          {
            "in": "path",
            "name": "authorId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the author to unsubscribe from"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully unsubscribed (no content)"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "articles": {
        "type": "object",
        "required": [
          "_id",
          "title",
          "article",
          "img",
          "date",
          "ownerId"
        ],
        "properties": {
          "_id": {
            "description": "Article unique ID (MongoDB ObjectId)",
            "type": "string",
            "example": "686809896ac14799aa90b54a"
          },
          "ownerId": {
            "description": "ID of the user who owns this article (MongoDB ObjectId)",
            "type": "string",
            "example": "686809896ac14799aa90b54a"
          },
          "title": {
            "description": "Article title",
            "type": "string",
            "example": "When anxiety fills the space: the path to peace"
          },
          "article": {
            "description": "Contents of the article",
            "type": "string",
            "example": "In today`s world, where the rapid pace of events and the endless flow of information often rob us of our inner peace, finding harmony in everyday life becomes a real challenge"
          },
          "img": {
            "description": "Link to the uploaded article image",
            "type": "string",
            "format": "uri",
            "example": "https://ftp.goit.study/img/harmoniq/772a8103927c4f30b8d8c48avq.jpg"
          },
          "date": {
            "description": "Date of creation of the article",
            "type": "string",
            "format": "date",
            "example": "2025-05-05"
          },
          "rate": {
            "description": "Popularity of the article",
            "type": "integer",
            "example": 2
          },
          "desc": {
            "description": "Short description of the article",
            "type": "string",
            "example": "Meditations that help restore inner peace"
          }
        }
      },
      "user": {
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "properties": {
          "_id": {
            "description": "User's id",
            "type": "string"
          },
          "name": {
            "description": "User's name",
            "type": "string"
          },
          "email": {
            "description": "User's email address",
            "type": "string"
          },
          "avatarUrl": {
            "description": "URL to user's avatar",
            "type": "string"
          },
          "articlesAmount": {
            "description": "Number of articles saved by the user",
            "type": "number"
          },
          "savedArticles": {
            "description": "List of IDs of articles saved by the user",
            "type": "array",
            "items": {
              "type": "string",
              "example": "64b2f9d8a1234567890abcde"
            }
          },
          "subscriptions": {
            "description": "List of IDs of users that the user is subscribed to",
            "type": "array",
            "items": {
              "type": "string",
              "example": "64b2f9d8a1234567890abcde"
            }
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "BadRequestError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message",
                    "errors"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "message",
                          "path",
                          "type",
                          "context"
                        ],
                        "properties": {
                          "message": {
                            "type": "string",
                            "example": "Title is required"
                          },
                          "path": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            },
                            "example": [
                              "title"
                            ]
                          },
                          "type": {
                            "type": "string",
                            "example": "any.required"
                          },
                          "context": {
                            "type": "object",
                            "required": [
                              "label",
                              "key"
                            ],
                            "properties": {
                              "label": {
                                "type": "string",
                                "example": "title"
                              },
                              "key": {
                                "type": "string",
                                "example": "title"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Article not found"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "InternalServerError"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Something went wrong on the server"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}